create new trainer module and move train file there
